controller: read wireless settings from UCI
https://jira.prplfoundation.org/browse/PPM-836

The prplMesh controller needs to be configured with the SSIDs and
credentials that have to be configured on the agents. Even though NBAPI
exists to configure this, there is a lot of existing software out there
that doesn't use it. Therefore, prplMesh should also read the
configuration out of the legacy wireless settings.

Modify the `main()` in controller to read wireless settings into
controller's database. If these settings are not overriden using NBAPI,
then the UCI settings will be used instead to configure the whole
network.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>