agent: ap_manager: replace Socket* with int
https://jira.prplfoundation.org/browse/PPM-966

This is a preparatory commit.

Currently, ap_manager uses `Socket *` as a handle to socket
connections. Next commits will replace this with a file descriptor
which is an `int`.

To prepare for that change, rewrite method that gets executed whenever a
new CMDU message is received and create a new method to send CMDU
messages. These new methods use an `int` instead of a `Socket *` so they
can be used by the new sockets API.

Temporary methods and member variables created for this purpose will
be removed at the end of PPM-966.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>