bwl: nl80211: avoid endless loop in generate_connected_clients_events
Method `generate_connected_clients_events()` is not implemented yet in
the NL80211 flavor of the BWL, it just prints a message and returns.

AP manager is calling this method on each FSM polling cycle so the log
is flooded with those "not implemented" messages.

To avoid it, set the value of the output parameter
`is_finished_all_clients` to true so the AP manager calls this method
only once. Final implementation will set this parameter appropriately.

Signed-off-by: mmazzapater <mmazzapater@gmail.com>