utils: os: extend system_call with retrieved output
system_call is designed to run commands.

It lacked of reading output of the command.
There is piece of code but commented out.

To solve this, reading output capability is added with popen().
Size of buffer (10k) limits maximum readability.
Also, output level can be set with desired output.

And also, error conditions are not handled well in ap_wlan_hal_dwpal.cpp.
But iwpriv commands are only exits for specific platforms (UGW, not CGR).
So, they always return true. It should be considered to fix later on.

Log level parameter is removed.
Because log levels should be system wide.

https://jira.prplfoundation.org/browse/PPM-535
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>