agent: Recalculate byte units to support R2 spec
R2 specification require associated station traffic stats to be in KB
units while R1 only support Byte units.
The monitor produces the value in Byte units so it need to be
recalculated according to the controller expectation.

This commits handles the DB setting of the expected byte units in the
capability report and adds recalculation of the bytes according of the
reported format.

Signed-off-by: Semyon Barenboym <simonbaren@gmail.com>