agent: monitor: refactor FSM method
https://jira.prplfoundation.org/browse/PPM-967

This is a preparatory commit.

The method implementing the FSM in monitor is called periodically by
the monitor thread. In a follow-up commit the monitor will not be a
thread anymore. Instead, it will be the event loop running in the
main thread who will periodically call such FSM method. The FSM method
must return a boolean to tell the event loop if execution must
terminate.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>