agent: remove error notification from monitor to slave
https://jira.prplfoundation.org/browse/PPM-967

When an error occurs in `monitor_thread`, it sends
`ACTION_MONITOR_ERROR_NOTIFICATION` message to `slave_thread` and then
waits for ACK.

The design and implementation of this error notification mechanism is
flawed and does't work as expected:
- One of the errors that `monitor_thread` tries to notify to
`slave_thread` is when socket connection to slave gets unexpectedly
disconnected. There's no connection to notify such an error!
- `stop_monitor_thread()` ends up setting `should_stop` to true, so
thread is stopped and ACK will never arrive.
- If an error notification is sent in `stop_monitor_thread()`, then
resource deallocation is skipped.

Instead of fixing this specific flow and because handling the error
consists of just printing to log, remove it for now. Note however that
the slave will not be able to know the reason the monitor unexpectedly
disconnected. If we need it in the future we will add it back.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>