agent: ap_manager: create shared_ptr instead of unique_ptr
https://jira.prplfoundation.org/browse/PPM-967

This is a preparatory commit.

Currently, method `create_cmdu_client_factory` returns a `unique_ptr`
which is moved to the AP manager.

In fronthaul process, both AP manager and monitor need a CMDU client
factory to create clients to connect to CMDU server running in slave.

Instead of creating two factory objects, refactor the constructor of AP
manager to accept a `shared_ptr`. In the following commit, the same
factory object used by AP manager will be shared with monitor too.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>