bcl: reset fd on remove_timer()
Method `remove_timer()` cancels the timer identified by the given file
descriptor and removes the event handlers that were installed in event
loop when the timer was created. On method return, the file descriptor
given as parameter is not valid anymore. In many places of the code we
usually reset its value to mean that the timer has already been removed.

In order to make it explicit that the file descriptor is no longer valid
and to avoid that the calling party had to reset the file descriptor
after removing a timer, modify `remove_timer()` so the file descriptor
parameter is passed by reference and set it to `invalid_descriptor`
before returning.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>