tests: boardfarm: Add support for a remote linux station
After making the bridge between the actual device
and the docker controller, an IP on the same subnet
would still exist on the to-dut interface. This caused
the host to have a route preference to the station
through to-dut instead of the docker bridge interface.
This would also cause the wifi traffic on the remote
device to be returned from the wired interface, since they
were both on the same subnet and the wired one would have
priority.
To avoid this, a move_iface_ip_to_bridge function was
created to add a different subnet IP on the bridged
interface.
A new netgear-rax40-remote config was added because
otherwise the current test-bed would break when trying
to initalize the remote station device.
The config file assumes that the station_ip_wifi is already
static on the remote device wifi interface.

Signed-off-by: David <davidmgafonso@gmail.com>