controller: db: get_radio: return null if node is not slave
During refactoring of node hierarchy a parallel `sAgent/sRadio`
structure was created. During transitional period both
hierarhies will co-exist.

If a node changes its type [1], the two structures will be
inconsistent.

To prevent this, just pretend that a radio does not exist
if the node with the same mac address if not TYPE_SLAVE.

[1]: https://jira.prplfoundation.org/browse/PPM-723

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>