agent: Handle lack of traffic separation configuration TLV more correctly
According to the spec, if a Multi-AP Agent has not received a Multi-AP
Default 802.1Q Setting subelement in the (Re-)Association Response
frame, or a Traffic Separation Policy TLV, or the most recently received
Traffic Separation Policy TLV has the Number of SSIDs field set to zero,
the Multi-AP Agent shall consider the Primary VLAN ID not configured and
may forward VLAN tagged packets received at the Fronthaul interface or
Logical Ethernet interface, and shall not add, modify or remove any VLAN
tag on any packets it sends or receives and shall not perform traffic
separation on the basis of any VLAN tag present in a packet.

Therefore, remove the traffic separation configuration when not
configured, only when trafic seperation TLVs are not present or does not
contain any SSID VID mapping only on the MULTI_AP_POLICY_CONFIG_REQUEST
message handler and not on the M2 handler.

While at it, move the initialization of some AgentDB fields to the
backhaul INIT state instead of the RESTART state. This is a more appropriate
place to initialize it.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>