controller: db: return sAgent in add_node_ire/gateway
Return the `sAgent` object in `add_node_ire` and
`add_node_gateway`.

These methods are not used in boolean context, so
change API to return `sAgent` (a `true` value)
even in case of errors (a false positive error was
returned if a node with the given AL MAC already
existed).

PPM-1057

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>