tests: LinkMetricQuery: check for zero MAC addresses
While working on PPM-230, a bug lead to the MAC address of the wired
backhaul interface to be empty. It was only noticed by chance because
the NbAPIDevice test was failing on Tx/Rx packets not increasing. The
bug itself was discovered because some MAC addresses in the link
metric responses were empty. To catch it sooner in the future, check
the MAC addresses in the LinkMetricQuery test.

The MAC address of the device that transmits the link metric
response (agent2) and the one of the receiver (controller) were
already checked.

What we need to additionally check is the MAC addresses of the
interface pair between the two devices. It's a little bit of work to
check the actual values of the pair, so for now just check that they
are not zero.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>