platform_manager: fail if bridge_iface or wan.iface are empty
If the platform manager doesn't have values for bridge_iface or
wan.iface, it can break in subtle ways (for example: no backhaul
interface MAC in link metric responses).

Since we anyway require those two parameters to be set, fail early if
they are empty to make it easier to notice that there is a bug.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>