bcl: ignore unknown configuration lines
This is a preparatory commit.

When support for a configuration parameter is removed from BCL, it can
take some time until it's actually removed from every platform. For
example, UTAF tests on RDK-B will fail until the parameters are
removed (which can take some time, as it needs to be tested
beforehand).

Instead of failing to start, when an unknown configuration line is
found, print a warning on stdout and continue.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>