dev: boardfarm: removed hardcoded interfaces
ALEntityPrplWrt used to hard code the radio interface names to wlan0
and wlan2. In reality, devices can have about any interface names
(for example, (wlan0-1) is the default one on a non-MaxLinear prplWrt
device when one station interface is defined).
For MaxLinear devices, we want to use the first bss as the radio
interface, because it's a special, dummy one (for other devices it
doesn't matter).
The interface names are not mandatory in UCI, so we cannot use that
to find them.
Instead first query ubus to get the interface names for all radios.
Keep only the AP interfaces, then use one for each radio to ask hostapd
what its first bss is.
Also took a change to update get_iface function to handle names other
than ones started with "wlan" and modify get_prplMesh_status.

Signed-off-by: David <davidmgafonso@gmail.com>