bwl: multiple fBSS WPS adjustments
Currently, we are assigning some fBSS configurations used for WPS
on each fBSS configuration.
This is wrong, so add these configurations only on a single fBSS.
In addition add configurations that are missing on the fBSS that
perform WPS, to perform the WPS correctly.

While at it, make sure we are removing old configuration that are
not relevan to a bSS:
* max_num_sta
* num_vrt_bkh_netdevs
* multi_ap_profile1_disallow
* multi_ap_profile2_disallow

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>