tests: capi.py: add support for read timeouts
For the certification tests, we want to try to get a reply until a
timeout is reached.

The socket itself already has a timeout, but in some specific cases
the connection is kept in the recv() loop.

Add a `timeout` parameter to get_reply() that aborts if the timeout is
reached.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>