wlan_hal_dwpal: remove NL header offset from returned reply
When the wlan-hal-dwpal sends an NL request using the DWPAL it expects
the reply to have an NL_ATTR_HDR with the size of 4 bytes.

The above-mentioned NL_ATTR_HDR is now removed by the dwpal before the
reply is returned to the caller (in this case the prplmesh
wlan-hal-dwpal) and thus we don't need to consider it when calculating
reply size or copying the information from the reply buffer.

Removed the NL_ATTR_HDR from the calculation of NL reply size.
Copy and read of NL reply starts at the beginning of the buffer (and
not in an offset of NL_ATTR_HDR).
Removed the definition of the NL_ATTR_HDR.

Signed-off-by: Adam Dov <adov@maxlinear.com>