boardfarm: fix client association control message
When boardfarm tests client association it sends a Client Association
Control message with a Client Association Control TLV.
One of the fields of the TLV is Validity Period that is defined as a
uint16_t. However, the test formatted this value as uint8_t which i turn
caused an error message in the agent logs:
`tlvClientAssociationControlRequest.cpp[56] --> TLVF: Requested index
is greater than the number of available entries`

The test didn't fail on it since the son_slave doesn't check if the
value of the mac is valid.
This commit fixes the length of the field and the TLV.

Signed-off-by: Semyon Barenboym <simonbaren@gmail.com>