agent: Add function fill_channel_list_to_agent_db()
Add a new function "fill_channel_list_to_agent_db()" receive channels list in
a tlv format, and save it on the AgentDB in the DB format.
Basically, this commit moves an already exist code that does it into a function.
This function will be used from several places in the future.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>