bcl: Expose has_operating_class_channel()
Make has_operating_class_channel() function on wireless_utils a class static file
(currently the fucntion is static on the cpp file).
The reason is that the function will be used externally in a future commit.
Also, change the input argument from sWifiChannel to an explicit channel
and bandwidth since it is more convenient when working with the standard TLVs.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>