tests: environment.py: docker: make logfilenames an instance attribute
This is a preparatory commit.

In a follow-up commit, checkpoint() will be implemented for log
files. To be able to do that, we need to know the list of log files
that an entity or radio uses.

Since the files are anyway always the same, construct them in each
class' constructor, instead of every time _docker_wait_for_log is
called. This will allow us to re-use the file names later on.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>