tests: factor out dev_reset_default
This is a preparatory commit.

We usually don't want to check for log lines that appear before
dev_reset_default was called. Factor it out in a dedicated method, to
make it easier to add a log checkpoint when it's called.

The DevResetDefault and CapiWirelessOnboarding tests are not updated
here, as they apply to the agent and not the controller.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>