tests: InitialApConfig: configure the agent as part of the test
The test previously didn't configure anything on its own, which means
it relied on the log files to contain the M2 log lines from any time
before. This is going to be a problem once checkpoint() is implemented
for log files.

Configure the agent as part of the test so that the lines
that are checked are actually triggered by the test.

The timeout for cmd_reply is set to 60 seconds because as can be seen
in [1], the reset can take up to 30 seconds on an actual device.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1319220063

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>