tests: client_steering_dummy: remove unwanted checkpoint() call
This checkpoint is problematic: it is placed after the steering
command, but before the logging lines for the steering are checked. We
could move the checks for the steering logging before the checkpoint
and keep the checkpoint for when the station actually steered, but
then there is still a race condition: with a dummy station we know
from the tests when it steers (since we steer it manually) but with an
actual station it can steer on it's own, so we have no way to know
when the checkpoint should happen.

Remove the checkpoint to treat the steering command and the actual
steering as a single "operation".

This was not a problem before because checkpoint() was not implemented
for logs yet. Since we want to implement it, we have to fix it first.

Note that it's only possible to remove the checkpoint because
check_topology_notification() doesn't fail anymore if multiple
topology notifications are found with different association events.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>