ApConfigBSSTeardown: wait until the connection map is updated
As described in PPM-1410, there is a timing issue when the connection
map is updated after an autoconfig renew. If we don't wait for long
enough, the radios might not be in the connection map because they are
not "active".

Instead of increasing the delay before we check the connection map,
add a logging line in the controller DB so that we have something we
can rely on to know that the connection map was updated. This allows
us to use a timeout before continuing the test instead of a fixed
delay (it reduces the time we have to wait for, and it is more
reliable).

Closes PPM-1410

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>