tests: environment.py: return strings in command()
For command() we had a mixed situation: command() on the prplWrt
devices used decode() to decode the output of the command and return
strings, while command() in docker didn't use it and returned bytes.

It's unlikely that we would want to treat the output of a command as
bytes and not strings, so take the "device" command() approach. Use
decode() in command() itself instead of in the callers.

Adapt the type hints, the callers, and the tests that use bytes
instead of strings.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>