tests: check_cmdu_type: raise instead of using assert
There might be cases where we want to check if a tlv is part of a
cmdu, but not fail if it's not. By raising an exception instead of
using assert, we can catch it in the caller and act accordingly.

In a follow-up commit this will be used to check for association event
tlvs in multicast topology notifications.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>