tests: check_topology_notification: allow multiple topology notifications
Previously check_topology_notification() failed if multiple topology
notifications were present in the capture, no matter what their
association event TLV was.

This can be a problem, because in some cases (example: a station being
steered) we can get multiple topology notifications since the last
checkpoint, but we want to allow it so long as their association event
is different.

Instead of checking that there is one and only one multicast topology
notification get all of them, and filter out the ones that do not
match the association event passed as an argument. If there is no or
multiple matching topology notifications, fail like it was the case
before.

We will take advantage of this new behavior in a follow-up commit, to
change the way the checkpoints are used in ClientSteeringDummy.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>