bcl: Add function get_overlapping_beacon_channels()
Add function get_overlapping_beacon_channels() to wireless_utils.
This function returns a list of all overlapping beacon channels for
a given channel and bandwidth.
The purpose of this function is to help fill the preference report
the Agent sends to the controller by marking a channel as non-operable
because one of its overlapping beacon channels is a DFS unavailable.
This function will be used on a future commit.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>