bcl: Add function center_channel_5g_to_beacon_channels()
Add function center_channel_5g_to_beacon_channels() to wireless_utils.
This function returns a list of all overlapping beacon channels for a given
center channel and bandwidth.
The purpose of this function is to help fill the preference report the
Agent sends to
the controller when filling the preference on a channel which in
operating classes 128, 129, 130 which use center channels instead of
beacon channels.
This function will be used on a future commit.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>