bcl: Move 3 wireless utils function to the agent
Move 3 wireless_utils functions to the agent tlv_utils.
These functions are only used to fill the AP Radio Basic Capabilities TLV.
The reason for doing this:
1. These functions are only used on tlv_utils on the Agent, so it is better
to move them to where they are exclusively used.

2. When replacing the preferred and supported channels list with the
unified channels list, we will have to use struct defined on the BTLVF lib,
which is dependent on BCL, therefore, we cannot keep it in the wireless utils.
As part of this change the function signature to use the new channels list.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>