controller: db: Skip data model for persistent clients
When loading clients from the persistent DB, which have not yet
connected, we need to skip adding them to the data-model DB.

This needs to be done for two reasons,
1) the loaded persistent client does not have a parent BSSID and thus
simply cannot be added to the data-model DB.
2) While the node DB can represent disconnected clients, the data-model
only only reflects connected clients and thus shouldn't contain the
persistent clients while they are not connected.

To resolve this we can skip the insertion to the data-model database if
a new client has no parent BSSID but only if the persistent DB is
enabled.

It is importent to validate that this happens only when the persistent
DB is enabled, because that is the only flow which should add a new
client that has no parent BSSID.

Resolves PPM-1413

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>