controller: db: refactor get_hostap_ssid to get vap SSID
This change reverts the change introduced by commit
edef746180c4fb77969ecb0750c173c335ad39f6 which mistakenly changed the
get_hostap_ssid (which reads vap information) to read the information
from the radio.

Renamed the API to get_hostap_vap_ssid() and renamed function parameter
to bssid for it to be more apparent that it should be used with bssid.

Changed the API implementation to read the ssid from the parent radio
in agent db and use the bssid to find the correct vap in that radio.

Signed-off-by: Adam Dov <adov@maxlinear.com>