agent: monitor: do not stop the event loop on error with nl events
In some cases on DWPAL-based devices, an EPOLLERR event is received on
the file descriptor that is used by the monitor to listen to nl
events.

When this happens, the event loop stops as described in PPM-1421,
which means the ap manager also completely stops.

We don't know what causes the error event to be received. To be able
to make progress on PPM-1392, for now ignore the problem by only
preventing the event loop from stopping when the error occurs.

Note that this means the monitor will not receive nl events from the
point where the error was received onwards.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>