bpl: skip configuration of hidden BSS when reading wireless settings
MaxLinear devices have a dummy VAP for each radio, which should not be
used when reading the existing wireless settings.

Besides the fact that they miss the 'encryption' option (which is
optional in OpenWrt) the only thing that we can use to distinguish
them from regular VAPs is the 'hidden' option.

Aside from that, prplMesh's representation of a BSS configuration
anyway doesn't support the hidden option, and we most likely don't
want the controller to propagate its hidden SSIDs to other agents.

When a hidden BSS is found, log it and skip it.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>