bcl: Fix function get_5g_center_channel()
The function "get_5g_center_channel()" is depended on the function
"channel_to_vht_center_freq()" which is known for not always return
a correct retuyrn value.
Also, it has a third argument "channel_ext_above_secondary" which
is not filled correctly.
Replace the function implementation to be based on the 5G channel
table instead.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>