bcl: Fix fucntion has_operating_class_channel()
The function has_operating_class_channel() checks if a channel
exists on an operating class table entry.
In case the entry is from operating classes 128/129/130 which
use a center channel, this function will return a wrong return
value since it does not consider that.
Fix the function so it will handle properly entry from the operating
classes mentioned above.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>