agent: Request for channels list after the AP Manager joins
Since from this MR the Agent is now relies only on the channels list and
not the old preferred/supported channels list, we need to make sure that
the Agent has the channels list.
Add a request for channels list after the AP manager joins.
This change is temporary and will be removed on the next MR which will
send the channels list on the Join Message.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>