controller: remove al_mac from send renew message
According to the 'IEEE STD 1905.1-2013' the AP_AUTOCONFIG_RENEW_MESSAGE
should be sent using multicast therefore there is no need to send an
al_mac argument to the function.

In controller::access_point_commit need to replace the for loop with a
condition.
If there is atleast one IRE present send a multicast RENEW message.

This is a preparative commit

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>