controller: send renew on agent monitoring task
Usually, the ap-configuration sequence is started when the agent sends
the ap-config search request message, to which the controller responds
with the ap-config response message.

But with the scenario presented in PPM-1390, the controller process is
restarted and loads without ever receiving an ap-config search request
message.

To resolve this, we can initialize the ap-config "handshake" by sending
a ap-config renew message, informing the agent of a newly restarted
controller process.

To make sure all the tasks in the controller are running currectly, we
need to send the ap-config renew message after the controller finished
its startup function, during the event loop handling.
Doing this will give all the controller tasks enough time to finish
their initialization.

Since the agent_monitoring_task is suppoed to utilize the AGENT_JOIN
handling, it was suggested as the best place to send the ap-config
renew message from.

Resolves PPM-1390

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>