controller: set prplmesh on SLAVE_JOINED_NOTIFICATION
When an agent sends a ACTION_CONTROL_SLAVE_JOINED_NOTIFICATION, update
the DB and mark the agent as prplmesh.

While this is done during the AP Autoconfiguration Search message
handling, But with the scenario presented in PPM-1390, the controller
process is restarted and loads without ever receiving an ap-config
search request message thus is unaware that the agent is prplmesh.

Resolves PPM-1390

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>