bpl: skip configuration with empty SSID when reading wireless settings
When the SSID of a BSS is not set in UCI, hostapd will not be
configured correctly:
Tue Jun 15 05:35:59 2021 daemon.err hostapd: Line 33: invalid SSID ''

It's important that the controller doesn't propagate this
configuration to the agents, as it will cause the agents to tear down
their radio(s).

When the controller reads a BSS with an empty or unset SSID, skip it.

While at it:

- Move the check for the "mode" further up, so that we don't try to
look for parameters like `disabled`, `hidden` or the SSID if the mode
is not `ap`.

- Take the chance to update the comment about the mode, since the mode
doesn't indicate fronthaul/backhaul, but rather whether it's an AP or
something else (STA, adhoc, etc).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>