bwl: Add missing ACTION_BACKHAUL_CONNECTED_NOTIFICATION
Add sACTION_BACKHAUL_CONNECTED_NOTIFICATION which contain the Multi-AP
profile, and the Primary VLAN ID. Since this is not implemented on this flavour of BWL,
leave it empty.

Lack of it cause a crash on the backhaul manager that derefernce a pointer to this message.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>
(cherry picked from commit 9905db1f9810e4cc6ce2d8ec63672669a3f6e3de)