agnet: Move get_channel_preferences() to son_slave
The function get_channel_preferences() on wireless_utils gets a list of supported
channels and returns a channels preference list which is used for reporting the
channel preference to the Controller.

Move this function to the son_slave, and change it to receive the channels list from
the AgentDB instead.

The main reason for the function moving is since the channels list is depended on the
btlvf which the bcl cannot be depended on.
Since this function is only used on the son_slave it is okay.

As part of this commit create a new channel preference struct which will allow to arrange
the channels preference more easily for send it to the Controller.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>