agent: Send Channel Preference Report to the Controller
When the son_slaves receives the CHANNELS_LIST_RESPONSE message from the
AP Manager, it generates and sends CHANNEL_PREFERENCE_REPORT_MESSAGE
to the Controller.

This commit removes the saving of the preferred channels on the message to the
AgentDB, and also use  the new function "get_channel_preferences_from_channels_list()"
from the preveious commit to build the message.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>