agent: Refactor channel_selection_get_channel_preference()
The function "channel_selection_get_channel_preference()" is repsonsible to save the
channel prerefence sent by the controller on CHANNEL_SELECTION_REQUEST_MESSAGE.

Rename the function to "get_controller_channel_preference" which is a more intuitive name.
Also, refactor the function to save the preference on the new channel preference container.
Besides, removes redundant casting in log prints on the function.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>