tests: increase socket timeout for prplWrt devices
As can be seen in [1], the InitialApConfig test can still
occasionnally fail on the rax40.

By looking at the log files, we can see the dev_reset_default command
takes around 30 seconds to complete.

Commit b1cfe98f9ba82a5dbd6dc78a425756b7c722d40e was supposed to
address it by increasing the timeout given to cmd_reply() to 60
seconds.

As we can see in the logs, in turns out the timeout that is reached is
the internal timeout of the socket (which defaults to 30 seconds), not
the one used by cmd_reply(). To fix it, set the socket timeout of the
UCCSocket to 60 seconds for prplWrt devices.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1355236729

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>