agent: Fix wrong bandwidth
On the CHANNEL_SELECTION_REQUEST_MESSAGE handler, we check
if we need to switch to a different channel.
If we don't need to change the channel, we send a channel switch
request to the ap manager with our current channel.
We do it since we only need to change the tx power limit, without
actual channel change.

The issue is that the bandwidth value is set with the uninitialized
variable "channel_to_switch".
Change the value of the bandwidth to our current bandwidth in
a case when we don't need to switch the channel.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>