agent: Compare also the bandwidth
When the AP manager receives a channel switch request, it checks if the current
channel is identical to the requested channel. If it does, it doesn't switch the channel
and send back a default reply.
Checking only the channel is not enough. Need to compare the bandwidth as well.
Add a comparison of the bandwidth in addition to the channel, to decide whether to
call channel switch or not.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>