tests: boardfarm: ApConfigBSSTeardownCli: replace sleep with check_log()
As can be seen in [1] and [2], ApConfigBSSTeardownCli has the same
problem than the one that was addressed for ApConfigBSSTeardown in
`b8308e7e7b3d2a7b242c84e388fa6a6263bcb4a9`: the sleep(3) statement is
sometimes not enough for the connection map to update completely.

Fix it in the same way: use check_log() to wait for the connection map
to update.

The first sleep(3) doesn't use the connection map, instead it checks
the vaps on the agents directly. However waiting for the connection
map to update should make sure the agents themselves were updated.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1367664488
[2]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1368133865

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>