agent: Comment out misconfigure BSS tear down
We currently do not support bBSS with both profiles disallow flags set to false.
When we are configured in a way we don't support, we should tear down the BSS,
andsend an error response on that BSS.
Since the Certification tests have a bug that ends up configuring the Agent wrongly.
To deal with it, temporarily comment the code snippet generates an error on that
BSS and tear it down, to allow it to be configured successfully.
When the issue with the testbed will be fixed, comment out the line above.

In addition, remove the number of configured VLANs as a condition to the bad configuration check since it is irrelevant.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>