controller: db: add get_station method
Add `get_station` method that looks for a station based on
its MAC address.

An alternative approach would be to use `m_stations` variable
directly, as it is done with `sAgent` objects. Since it is
not yet clear whether stations should be stored in one
singe map (as agents) in the DB class or stored in separate
maps in `sBSS` objects, access station through `get_station`.
In case the decision changes, only `get_station` will need
to be rewritten.

PPM-1058

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>