controller: m1 handler: create bridge object in advance
There is some overlap in the code handling joinings of
prplmesh and non-prplmesh devices.

Non-prplmesh devices always join but prplmesh ones are
rejected in some circumstances.
It prevents from just moving shared code into M1 handler.

As a workaround, create the `sAgent` object representing
the bridge in advance, but put it into a disconnected state.

Checks for the connected state are used in many places of
the controller code anyway.

This change also splits the responsibilities. There is no need
to connect a device just to store some information about it
in the DB.

PPM-1057

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>