tree-wide: remove unused STL containers headers
A script (see below) was used to find potential
candidates for the removal. It abuses the fact
that an STL container and its header file often
have the same name.

This script does not modify files, it prints
a list of suggestions.

containers="array vector deque forward_list list \
            stack queue priority_queue set map \
            unordered_set unordered_map"

for container in $containers; do
    echo
    echo ==== $container ====
    echo
    grep -rnIH include..$container \
        | grep -v ^build \
        | cut -d : -f 1 \
        | xargs grep -nH $container \
        | cut -d : -f 1 \
        | uniq -c \
        | grep ' 1 ' \
        | sed -e 's/^.*1 //g' \
        | xargs grep --color -nH $container
done

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>