controller: db: do not remove agent in remove_node
`remove_node` is called only if a node with corresponding
MAC address exists but has a wrong type.

This usage is irrelevant to the new `sAgent/sRadio/sStation`
structure.

Moreover, it can be harmful in case of
`handle_(non)_intel_slave_join`.
It can delete an agent partially prepared in the M1 handler.

Change `remove_node` to not remove `sAgent` objects.

PPM-1057

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>