nbapi/odl: change uint8 and int8 to unit32 and int32 with limit
As ubus tool and libubox makes a json string and converts an int8 and uint8 to "true" or "false"
depending on if it is non-zero or zero it causes to falling of boardfarm tests.

Changed all used previously uint8 and int8 to more reliable uint32 and int32
with appropriate limitations of max and min values. 
Changed limits for MaxTxPower according to specification.

--
PPM-994
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>