agent: Refactor recalculate_byte_units method
The `recalculate_byte_units` was introduced with bytes parameter passed
by reference. Since it's not the best practice, refactor the method so
the bytes passed by value.
While at it, remove the special handling for 0 value since it is not
needed - the method does not do the division by the bytes value.

Signed-off-by: Semyon Barenboym <simonbaren@gmail.com>