controller: nbapi: on_action: remove unused headers
`<chrono>` header defines everything in `std::chrono`
namespace, so checking whether the header is used can
be done by looking for the word `chrono` in the file.

`<chrono>` is not used in `nbapi/on_action.h` but is
used in the cpp file. Move it there.

Checking whether `<ctime>` is used can be done by
looking for all things defined there as it is done
in [1].
The only used definition is `NULL`. But since `NULL`
is defined in many standard files it is fine to delete
this header.

[1]:
definitions="\
    CLOCKS_PER_SEC \
    TIME_UTC \
    NULL \
    clock_t \
    size_t \
    time_t \
    tm \
    timespec \
    clock \
    time \
    timespec_get \
    ctime \
    asctime \
    strftime \
    wcsftime \
    gmtime \
    localtime \
    mktime"

for def in $definitions; do
    grep '\b'$def'\b' controller/nbapi/on_action.*
done

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>