controller: db: remove unused TLV headers
TLV headers looks like this:
`#include <path/to/tlv_name.h>`
(`tlv_name` is a placeholder for any TLV header name).

If a string `tlv_name` is found in the file only once
(in the `#include` statement) it is a good indication that
the header is not used.

Looking for unused TLV include statements is a good
candidate for automation.

It is also possible to look for unused TLV headers manually
with a decent speed using `gedit`:
1) double click on `tlv_name` to select the whole name;
2) press `CTRL+F` to search for `tlv_name`;
3) check the number of matches;
4) if it is 1 - remove the `#include` statement.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>