tree-wide: remove unused #include <bcl/son/son_wireless_utils.h>
All methods of `son_wireless_utils.h` are located in
a namespace called `son::wireless_utils`.
It allows to create a simple script (the exact code has
been lost) to find unused inclusions of this file.

The main idea is that if a file contains
`#include <bcl/son/son_wireless_utils.h>` and does
not contain any other line that has `wireless_utils`,
this `#include` is redundant.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>